Pick Item function and Interaction
Added a Script to select a random item from a list at start and display it in the Requirements Tab
Logic: all items players can select in the scene should have the "itemsLayer" layer. The selected item at beginning of the scene gets the "validItemLayer" layer.
Added to the Ray script to check if the player is selecting the right or wrong item.
Current items used as examples are "picture 1" "picture 2" and "picture 3"